+++ app/defaults.py
@@ -1,0 +1,33 @@
+DEFAULT_BOOKMARK_LIST = [
+    {
+        "id": "0",
+        "href": "https://www.notion.so/",
+        "size": 48,
+        "title": "notion",
+        "imgUrl": "/images/icon-notion.svg",
+        "bgColor": "#fff",
+    },
+    {
+        "id": "1",
+        "href": "https://github.com/",
+        "size": 38,
+        "title": "github",
+        "imgUrl": "/images/icon-github.com.svg",
+    },
+    {
+        "id": "2",
+        "href": "https://juejin.cn/",
+        "size": 38,
+        "title": "掘金",
+        "imgUrl": "/images/icon-juejin.cn.svg",
+    },
+    {
+        "id": "3",
+        "href": "https://bilibili.com/",
+        "size": 48,
+        "title": "bilibili",
+        "imgUrl": "/images/icon-bilibili.svg",
+        "bgColor": "#fff",
+    },
+]
+